Reorganise functional system test suite
