+++ test/functional/system/blocks/chain.delete_last_block.js
@@ -23,7 +23,7 @@ var localCommon = require('../common');
 
 describe('system test (blocks) - chain/deleteLastBlock', () => {
 	var library;
-	localCommon.beforeBlock('system_blocks_chain', lib => {
+	localCommon.beforeBlock('system_blocks_chain_delete_last_block', lib => {
 		library = lib;
 	});
