+++ test/functional/system/blocks/process/process.js
@@ -16,15 +16,15 @@
 'use strict';
 
 var async = require('async');
-var constants = require('../../../../helpers/constants');
-var genesisBlock = require('../../../data/genesis_block.json');
-var application = require('../../../common/application');
-var modulesLoader = require('../../../common/modules_loader');
-var clearDatabaseTable = require('../../../common/db_sandbox')
+var constants = require('../../../../../helpers/constants');
+var genesisBlock = require('../../../../data/genesis_block.json');
+var application = require('../../../../common/application');
+var modulesLoader = require('../../../../common/modules_loader');
+var clearDatabaseTable = require('../../../../common/db_sandbox')
 	.clearDatabaseTable;
 var loadTables = require('./process_tables_data.json');
 
-describe('blocks/process', () => {
+describe('system test (blocks) - process', () => {
 	var blocksProcess;
 	var blocks;
 	var db;
@@ -36,7 +36,7 @@ describe('blocks/process', () => {
 		originalBlockRewardsOffset = constants.rewards.offset;
 		constants.rewards.offset = 150;
 		application.init(
-			{ sandbox: { name: 'lisk_test_blocks_process' } },
+			{ sandbox: { name: 'system_blocks_process' } },
 			(err, scopeInit) => {
 				blocksProcess = scopeInit.modules.blocks.process;
 				blocks = scopeInit.modules.blocks;
