+++ test/functional/system/transactions/2.delegates/4.different.accounts.different.usernames.js
@@ -15,14 +15,15 @@
 'use strict';
 
 var lisk = require('lisk-js');
-var accountFixtures = require('../../../fixtures/accounts');
-var randomUtil = require('../../../common/utils/random');
-var normalizer = require('../../../common/utils/normalizer');
-var localCommon = require('./../common');
+var async = require('async');
+var accountFixtures = require('../../../../fixtures/accounts');
+var randomUtil = require('../../../../common/utils/random');
+var normalizer = require('../../../../common/utils/normalizer');
+var localCommon = require('../../common');
 
 describe('system test (type 2) - double delegate registrations', () => {
 	var library;
-	localCommon.beforeBlock('system_2_2_delegates_2', lib => {
+	localCommon.beforeBlock('system_2_2_delegates_4', lib => {
 		library = lib;
 	});
 
@@ -32,10 +33,10 @@ describe('system test (type 2) - double delegate registrations', () => {
 	while (i < 30) {
 		describe('executing 30 times', () => {
 			var account = randomUtil.account();
+			var account2 = randomUtil.account();
 			var transaction;
 			var transaction1;
 			var transaction2;
-			var differentDelegateName = randomUtil.delegateName();
 			transaction = lisk.transaction.createTransaction(
 				account.address,
 				1000 * normalizer,
@@ -49,11 +50,29 @@ describe('system test (type 2) - double delegate registrations', () => {
 				});
 			});
 
-			describe('with same account using different usernames', () => {
+			describe('with two different accounts using different username', () => {
+				before(done => {
+					transaction1 = lisk.transaction.createTransaction(
+						account.address,
+						1000 * normalizer,
+						accountFixtures.genesis.password
+					);
+					transaction2 = lisk.transaction.createTransaction(
+						account2.address,
+						1000 * normalizer,
+						accountFixtures.genesis.password
+					);
+					localCommon.addTransactionsAndForge(
+						library,
+						[transaction1, transaction2],
+						done
+					);
+				});
+
 				it('adding to pool delegate registration should be ok', done => {
 					transaction1 = lisk.delegate.createDelegate(
 						account.password,
-						differentDelegateName
+						account.username
 					);
 					localCommon.addTransaction(library, transaction1, (err, res) => {
 						expect(res).to.equal(transaction1.id);
@@ -61,10 +80,10 @@ describe('system test (type 2) - double delegate registrations', () => {
 					});
 				});
 
-				it('adding to pool delegate registration from same account and different name should be ok', done => {
+				it('adding to pool delegate registration from different account and same username should be ok', done => {
 					transaction2 = lisk.delegate.createDelegate(
-						account.password,
-						account.username
+						account2.password,
+						account2.username
 					);
 					localCommon.addTransaction(library, transaction2, (err, res) => {
 						expect(res).to.equal(transaction2.id);
@@ -79,48 +98,42 @@ describe('system test (type 2) - double delegate registrations', () => {
 						});
 					});
 
-					it('first delegate registration to arrive should not be included', done => {
-						var filter = {
-							id: transaction1.id,
-						};
-						localCommon.getTransactionFromModule(
-							library,
-							filter,
-							(err, res) => {
-								expect(err).to.be.null;
-								expect(res)
-									.to.have.property('transactions')
-									.which.is.an('Array');
-								expect(res.transactions.length).to.equal(0);
+					it('both transactions should be included', done => {
+						async.every(
+							[transaction1, transaction2],
+							(transaction, callback) => {
+								var filter = {
+									id: transaction.id,
+								};
+
+								localCommon.getTransactionFromModule(
+									library,
+									filter,
+									(err, res) => {
+										expect(err).to.be.null;
+										expect(res)
+											.to.have.property('transactions')
+											.which.is.an('Array');
+										expect(res.transactions.length).to.equal(1);
+										expect(res.transactions[0].id).to.equal(transaction.id);
+										callback(null, !err);
+									}
+								);
+							},
+							() => {
 								done();
 							}
 						);
 					});
 
-					it('last delegate registration to arrive should be included', done => {
-						var filter = {
-							id: transaction2.id,
-						};
-						localCommon.getTransactionFromModule(
-							library,
-							filter,
-							(err, res) => {
-								expect(err).to.be.null;
-								expect(res)
-									.to.have.property('transactions')
-									.which.is.an('Array');
-								expect(res.transactions.length).to.equal(1);
-								expect(res.transactions[0].id).to.equal(transaction2.id);
-								done();
-							}
-						);
+					it('adding to pool delegate registration with already registered username should fail', done => {
+						localCommon.addTransaction(library, transaction1, err => {
+							expect(err).to.equal('Account is already a delegate');
+							done();
+						});
 					});
 
 					it('adding to pool delegate registration from same account should fail', done => {
-						transaction2 = lisk.delegate.createDelegate(
-							account.password,
-							randomUtil.delegateName()
-						);
 						localCommon.addTransaction(library, transaction2, err => {
 							expect(err).to.equal('Account is already a delegate');
 							done();
