+++ test/functional/system/transactions/4.multisignature/4.X.multisignature_unconfirmed.js
@@ -15,10 +15,10 @@
 'use strict';
 
 var lisk = require('lisk-js');
-var randomUtil = require('../../common/utils/random');
-var Scenarios = require('../common/scenarios');
-var transactionTypes = require('../../../helpers/transaction_types.js');
-var localCommon = require('./common');
+var randomUtil = require('../../../../common/utils/random');
+var Scenarios = require('../../../common/scenarios');
+var transactionTypes = require('../../../../../helpers/transaction_types.js');
+var localCommon = require('../../common');
 
 describe('system test (type 4) - sending transactions on top of unconfirmed multisignature registration', () => {
 	var library;
