+++ test/functional/system/transactions/5.5.dapps.js
@@ -16,10 +16,10 @@
 
 var lisk = require('lisk-js');
 var async = require('async');
-var accountFixtures = require('../../fixtures/accounts');
-var randomUtil = require('../../common/utils/random');
-var normalizer = require('../../common/utils/normalizer');
-var localCommon = require('./common');
+var accountFixtures = require('../../../fixtures/accounts');
+var randomUtil = require('../../../common/utils/random');
+var normalizer = require('../../../common/utils/normalizer');
+var localCommon = require('../common');
 
 describe('system test (type 5) - dapp registrations with repeated values', () => {
 	var library;
